requirements.txt
# Core web framework
streamlit>=1.48.0

# AI/LLM
openai>=1.99.0

# HTTP requests
requests>=2.32.0

# Environment management
python-dotenv>=1.0.0

# Alternative web framework (for simple_app.py)
flask>=3.1.0

# Optional: For better performance
watchdog>=4.0.0